+++ src/Stock_Package/Stock_Controller.java
@@ -35,7 +35,7 @@ public class Stock_Controller  implements Initializable {
     public TableColumn<Product,Integer> id_column;
     public TableColumn<Product,String> ref_column;
     public TableColumn<Product,String> des_column;
-    public TableColumn<Product,Integer> quantity_column;
+    public TableColumn<Product,Double> quantity_column;
     public TableColumn<Product,Integer> nbr_pcs_crt__column;
     public TableColumn<Product,Integer> nbr_pcs_column;
     public TableColumn<Product,Double> prix_vent_column;
@@ -94,7 +94,7 @@ public class Stock_Controller  implements Initializable {
             data = FXCollections.observableArrayList();
              rs = cnn.createStatement().executeQuery("SELECT  id,ref,des,nbr_pcs_crt,quan,nbr_pcs,prix_vent,value FROM demo.product_table");
             while(rs.next()){
-                data.add(new Product(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getInt(4),rs.getInt(5),rs.getInt(6),rs.getDouble(7),rs.getDouble(8)));
+                data.add(new Product(rs.getInt(1),rs.getString(2),rs.getString(3),rs.getInt(4),rs.getDouble(5),rs.getInt(6),rs.getDouble(7),rs.getDouble(8)));
             }
            }
         catch(SQLException eX){
